+++ Nos/Controller/PagedNoteDataSource.swift
@@ -11,6 +11,9 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
     var collectionView: UICollectionView
     
     @Dependency(\.relayService) private var relayService: RelayService
+    private(set) var databaseFilter: NSFetchRequest<Event>
+    private(set) var relayFilter: Filter
+    private(set) var relay: Relay?
     private var pager: PagedRelaySubscription?
     private var context: NSManagedObjectContext
     private var header: () -> Header
@@ -26,12 +29,14 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
     init(
         databaseFilter: NSFetchRequest<Event>, 
         relayFilter: Filter, 
+        relay: Relay?, 
         collectionView: UICollectionView, 
         context: NSManagedObjectContext,
         @ViewBuilder header: @escaping () -> Header,
         @ViewBuilder emptyPlaceholder: @escaping (@escaping () -> Void) -> EmptyPlaceholder,
         onRefresh: @escaping () -> NSFetchRequest<Event>
     ) {
+        self.databaseFilter = databaseFilter
         self.fetchedResultsController = NSFetchedResultsController<Event>(
             fetchRequest: databaseFilter,
             managedObjectContext: context,
@@ -40,6 +45,8 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
         )
         self.collectionView = collectionView
         self.context = context
+        self.relayFilter = relayFilter
+        self.relay = relay
         self.header = header
         self.emptyPlaceholder = emptyPlaceholder
         self.onRefresh = onRefresh
@@ -67,15 +74,23 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
             Log.error(error)
         }
         
-        Task {
-            var limitedFilter = relayFilter
+        subscribeToEvents(matching: relayFilter, from: relay)
+    }
+    
+    func subscribeToEvents(matching filter: Filter, from relay: Relay?) {
+        self.relayFilter = filter
+        self.relay = relay
+        
+        Task { 
+            var limitedFilter = filter
             limitedFilter.limit = pageSize
-            self.pager = await relayService.subscribeToPagedEvents(matching: limitedFilter)
+            self.pager = await relayService.subscribeToPagedEvents(matching: limitedFilter, from: relay?.addressURL)
             loadMoreIfNeeded(for: IndexPath(row: 0, section: 0))
         }
     }
     
     func updateFetchRequest(_ fetchRequest: NSFetchRequest<Event>) {
+        self.databaseFilter = fetchRequest
         self.fetchedResultsController = NSFetchedResultsController<Event>(
             fetchRequest: fetchRequest,
             managedObjectContext: context,
@@ -86,6 +101,7 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
         try? self.fetchedResultsController.performFetch()
         loadMoreIfNeeded(for: IndexPath(row: 0, section: 0))
         collectionView.reloadData()
+        collectionView.setContentOffset(.zero, animated: false)
     }
     
     // MARK: - UICollectionViewDataSource
@@ -208,7 +224,9 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
             startAggressivePaging()
             return
         } else if indexPath.row.isMultiple(of: pageSize / 2) {
-            Task { await pager?.loadMore() }
+            Task {
+                await pager?.loadMore()
+            }
         } 
     }
     
@@ -241,7 +259,9 @@ class PagedNoteDataSource<Header: View, EmptyPlaceholder: View>: NSObject, UICol
 
                 if self.largestLoadedRowIndex > lastPageStartIndex {
                     // we are still on the last page of results, keep loading
-                    Task { await self.pager?.loadMore() }
+                    Task {
+                        await self.pager?.loadMore()
+                    }
                 } else {
                     // we've loaded enough, go back to normal paging
                     self.stopAggressivePaging()