+++ Nos/Service/Relay/Filter.swift
@@ -5,13 +5,13 @@ import Foundation
 struct Filter: Hashable, Identifiable {
     
     /// List of author identifiers the Filter should be constrained to.
-    let authorKeys: [RawAuthorID]
+    var authorKeys: [RawAuthorID]
 
     /// List of event identifiers the Filter should be constrained to.
     let eventIDs: [RawEventID]
 
     /// List of Note kinds to filter
-    let kinds: [EventKind]
+    var kinds: [EventKind]
 
     /// An array of replaceable identifiers, or `"d"` tags, to match.
     let dTags: [RawReplaceableID]